+++ graphql_client/tests/operation_selection.rs
@@ -23,24 +23,6 @@ pub struct Heights;
 )]
 pub struct Echo;
 
-// The default is the first operation so this should be the same as Heights
-#[derive(GraphQLQuery)]
-#[graphql(
-    query_path = "tests/operation_selection/queries.graphql",
-    schema_path = "tests/operation_selection/schema.graphql",
-    response_derives = "Debug,PartialEq"
-)]
-pub struct Unrelated;
-
-#[derive(GraphQLQuery)]
-#[graphql(
-    query_path = "tests/operation_selection/queries.graphql",
-    schema_path = "tests/operation_selection/schema.graphql",
-    response_derives = "Debug,PartialEq",
-    selected_operation = "Echo"
-)]
-pub struct SelectedOperation;
-
 const HEIGHTS_RESPONSE: &str = r##"{"mountainHeight": 224, "buildingHeight": 12}"##;
 const ECHO_RESPONSE: &str = r##"{"echo": "tiramisù"}"##;
 
@@ -48,11 +30,7 @@ const ECHO_RESPONSE: &str = r##"{"echo": "tiramisù"}"##;
 fn operation_selection_works() {
     let heights_response_data: heights::ResponseData =
         serde_json::from_str(HEIGHTS_RESPONSE).unwrap();
-    let heights_unrelated_response_data: unrelated::ResponseData =
-        serde_json::from_str(HEIGHTS_RESPONSE).unwrap();
     let echo_response_data: echo::ResponseData = serde_json::from_str(ECHO_RESPONSE).unwrap();
-    let selected_operation_response_data: selected_operation::ResponseData =
-        serde_json::from_str(ECHO_RESPONSE).unwrap();
 
     let _echo_variables = echo::Variables {
         msg: Some("hi".to_string()),
@@ -61,14 +39,6 @@ fn operation_selection_works() {
     let _height_variables = heights::Variables {
         building_id: "12".to_string(),
         mountain_name: Some("canigou".to_string()),
-    };
-    let _unrelated_variables = unrelated::Variables {
-        building_id: "12".to_string(),
-        mountain_name: Some("canigou".to_string()),
-    };
-
-    let _selected_operation_variables = selected_operation::Variables {
-        msg: Some("hi".to_string()),
     };
 
     let expected_echo = echo::ResponseData {
@@ -80,22 +50,8 @@ fn operation_selection_works() {
         building_height: Some(12),
     };
 
-    let expected_heights_unrelated = unrelated::ResponseData {
-        mountain_height: Some(224),
-        building_height: Some(12),
-    };
-
-    let expected_selected_operation = selected_operation::ResponseData {
-        echo: Some("tiramisù".to_string()),
-    };
-
     assert_eq!(expected_echo, echo_response_data);
     assert_eq!(expected_heights, heights_response_data);
-    assert_eq!(expected_heights_unrelated, heights_unrelated_response_data);
-    assert_eq!(
-        expected_selected_operation,
-        selected_operation_response_data
-    );
 }
 
 #[test]
@@ -103,15 +59,8 @@ fn operation_name_is_correct() {
     let echo_variables = echo::Variables {
         msg: Some("hi".to_string()),
     };
-    let selected_operation_variables = selected_operation::Variables {
-        msg: Some("hi".to_string()),
-    };
 
     let height_variables = heights::Variables {
-        building_id: "12".to_string(),
-        mountain_name: Some("canigou".to_string()),
-    };
-    let unrelated_variables = unrelated::Variables {
         building_id: "12".to_string(),
         mountain_name: Some("canigou".to_string()),
     };
@@ -121,15 +70,5 @@ fn operation_name_is_correct() {
     assert_eq!(
         Heights::build_query(height_variables).operation_name,
         "Heights"
-    );
-
-    assert_eq!(
-        Unrelated::build_query(unrelated_variables).operation_name,
-        "Heights"
-    );
-
-    assert_eq!(
-        SelectedOperation::build_query(selected_operation_variables).operation_name,
-        "Echo"
     );
 }