+++ graphql_client_cli/src/generate.rs
@@ -1,47 +1,45 @@
 use failure;
 use graphql_client_codegen::{
-    deprecation, generate_module_token_stream, GraphQLClientCodegenOptions,
+    generate_module_token_stream, CodegenMode, GraphQLClientCodegenOptions,
 };
 use std::fs::File;
-use std::io::Write as IoWrite;
-use std::path::{Path, PathBuf};
+use std::io::Write as _;
+use std::path::PathBuf;
 use syn;
 
-#[allow(clippy::too_many_arguments)]
-pub fn generate_code(
-    query_path: PathBuf,
-    schema_path: &Path,
-    module_name: String,
-    selected_operation: Option<String>,
-    additional_derives: Option<String>,
-    deprecation_strategy: Option<&str>,
-    no_formatting: bool,
-    module_visibility: Option<&str>,
-    output: &Path,
-) -> Result<(), failure::Error> {
-    let deprecation_strategy = match deprecation_strategy {
-        Some("allow") => Some(deprecation::DeprecationStrategy::Allow),
-        Some("deny") => Some(deprecation::DeprecationStrategy::Deny),
-        Some("warn") => Some(deprecation::DeprecationStrategy::Warn),
-        _ => None,
-    };
+pub(crate) struct CliCodegenParams {
+    pub query_path: PathBuf,
+    pub schema_path: PathBuf,
+    pub selected_operation: Option<String>,
+    pub additional_derives: Option<String>,
+    pub deprecation_strategy: Option<String>,
+    pub no_formatting: bool,
+    pub module_visibility: Option<String>,
+    pub output_directory: Option<PathBuf>,
+}
 
-    let module_visibility = match module_visibility {
-        Some("pub") => syn::VisPublic {
+pub(crate) fn generate_code(params: CliCodegenParams) -> Result<(), failure::Error> {
+    let CliCodegenParams {
+        additional_derives,
+        deprecation_strategy,
+        no_formatting,
+        output_directory,
+        module_visibility: _module_visibility,
+        query_path,
+        schema_path,
+        selected_operation,
+    } = params;
+
+    let deprecation_strategy = deprecation_strategy.as_ref().and_then(|s| s.parse().ok());
+
+    let mut options = GraphQLClientCodegenOptions::new(CodegenMode::Cli);
+
+    options.set_module_visibility(
+        syn::VisPublic {
             pub_token: <Token![pub]>::default(),
         }
         .into(),
-        Some("private") => syn::Visibility::Inherited {},
-        _ => syn::VisPublic {
-            pub_token: <Token![pub]>::default(),
-        }
-        .into(),
-    };
-
-    let mut options = GraphQLClientCodegenOptions::new_default();
-
-    options.set_module_name(module_name);
-    options.set_module_visibility(module_visibility);
+    );
 
     if let Some(selected_operation) = selected_operation {
         options.set_operation_name(selected_operation);
@@ -55,7 +53,7 @@ pub fn generate_code(
         options.set_deprecation_strategy(deprecation_strategy);
     }
 
-    let gen = generate_module_token_stream(query_path, &schema_path, options)?;
+    let gen = generate_module_token_stream(query_path.clone(), &schema_path, options)?;
 
     let generated_code = gen.to_string();
     let generated_code = if cfg!(feature = "rustfmt") && !no_formatting {
@@ -64,8 +62,16 @@ pub fn generate_code(
         generated_code
     };
 
-    let mut file = File::create(output)?;
+    let query_file_name: ::std::ffi::OsString = query_path
+        .file_name()
+        .map(|s| s.to_owned())
+        .ok_or_else(|| format_err!("Failed to find a file name in the provided query path."))?;
 
+    let dest_file_path: PathBuf = output_directory
+        .map(|output_dir| output_dir.join(query_file_name).with_extension("rs"))
+        .unwrap_or_else(move || query_path.with_extension("rs"));
+
+    let mut file = File::create(dest_file_path)?;
     write!(file, "{}", generated_code)?;
 
     Ok(())